Added update callback to session service.
