+++ src/OpcUaStackClient/ServiceSet/SessionServiceStateId.h
@@ -1,0 +1,45 @@
+/*
+   Copyright 2019 Kai Huebl (kai@huebl-sgh.de)
+
+   Lizenziert gemäß Apache Licence Version 2.0 (die „Lizenz“); Nutzung dieser
+   Datei nur in Übereinstimmung mit der Lizenz erlaubt.
+   Eine Kopie der Lizenz erhalten Sie auf http://www.apache.org/licenses/LICENSE-2.0.
+
+   Sofern nicht gemäß geltendem Recht vorgeschrieben oder schriftlich vereinbart,
+   erfolgt die Bereitstellung der im Rahmen der Lizenz verbreiteten Software OHNE
+   GEWÄHR ODER VORBEHALTE – ganz gleich, ob ausdrücklich oder stillschweigend.
+
+   Informationen über die jeweiligen Bedingungen für Genehmigungen und Einschränkungen
+   im Rahmen der Lizenz finden Sie in der Lizenz.
+
+   Autor: Kai Huebl (kai@huebl-sgh.de)
+ */
+#ifndef __OpcUaStackClient_SessionServiceStateId_h__
+#define __OpcUaStackClient_SessionServiceStateId_h__
+
+#include "OpcUaStackCore/Base/os.h"
+
+namespace OpcUaStackClient
+{
+
+	enum class DLLEXPORT SessionServiceStateId
+	{
+		None,
+		Disconnected,
+		Connecting,
+		CreateSession,
+		ActivateSession,
+		Established,
+		Disconnecting,
+		Error
+	};
+
+	template<typename T>
+	std::ostream& operator<<(typename std::enable_if<std::is_enum<T>::value, std::ostream>::type& stream, const T& e)
+	{
+	    return stream << static_cast<typename std::underlying_type<T>::type>(e);
+	}
+
+}
+
+#endif
